% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs.R
\name{gs}
\alias{gs}
\title{Save ggplot2 object plot}
\usage{
gs(p, name = "gplot", outdir = base::getwd(), format = "pdf", w = 9, h = 9)
}
\arguments{
\item{p}{ggplot2 object plot}

\item{name}{output filename}

\item{outdir}{output directory}

\item{format}{output image format, pdf or png or both}

\item{w}{width}

\item{h}{height}
}
\value{
nothing
}
\description{
Save ggplot2 object plot
}
\examples{
gs(p = plot, outdir = "Result", name = "gplot", w = 9, h = 9)
}
