% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwd.R
\name{pwd}
\alias{pwd}
\title{Return the current working directory}
\usage{
pwd()
}
\value{
A character scalar containing the current working directory.
}
\description{
Return the absolute path of the current working directory.
}
\examples{
pwd()
}
