% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-scales.R
\name{canton_scales}
\alias{canton_scales}
\alias{scale_colour_canton_d}
\alias{scale_color_canton_d}
\alias{scale_fill_canton_d}
\alias{scale_colour_canton_c}
\alias{scale_color_canton_c}
\alias{scale_fill_canton_c}
\title{Canton colour scales for ggplot2}
\usage{
scale_colour_canton_d(palette = "NPG", ...)

scale_color_canton_d(palette = "NPG", ...)

scale_fill_canton_d(palette = "NPG", ...)

scale_colour_canton_c(palette = "NPG", ...)

scale_color_canton_c(palette = "NPG", ...)

scale_fill_canton_c(palette = "NPG", ...)
}
\arguments{
\item{palette}{Canton palette name.}

\item{...}{Additional arguments passed to the corresponding ggplot2 scale.}
}
\value{
A ggplot2 scale object.
}
\description{
Apply a palette returned by \code{hue()} to discrete or continuous ggplot2
colour and fill scales.
}
