% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs.R
\name{gs}
\alias{gs}
\title{Save ggplot2 object plot}
\usage{
gs(p, outdir = base::getwd(), name = "", w = 7, h = 7)
}
\arguments{
\item{p}{ggplot2 object plot}

\item{outdir}{output directory}

\item{name}{output filename}

\item{w}{width}

\item{h}{height}
}
\value{
nothing
}
\description{
Save ggplot2 object plot
}
\examples{
gs(p = plot, outdir = "Result", name = "A", w = 7, h = 7)
}
