% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwd.R
\name{pwd}
\alias{pwd}
\title{Print Current Working Directory}
\usage{
pwd()
}
\value{
a string of current working directory
}
\description{
Print the full filename of the current working directory.
}
\examples{
pwd()
}
