% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/using.R
\name{using}
\alias{using}
\title{Library multiple packages}
\usage{
using(...)
}
\arguments{
\item{...}{package names}
}
\value{
nothing.
}
\description{
Library multiple packages without starting messages.
}
\examples{
using(Seurat,tidyverse,magrittr,data.table)
}
