% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/using.R
\name{using}
\alias{using}
\title{Load Multiple Packages Without Starting Messages}
\usage{
using(...)
}
\arguments{
\item{...}{multiple package names without double quotation marks and seperated by comma}
}
\value{
nothing
}
\description{
Load Multiple Packages Without Starting Messages
}
\examples{
using(Seurat,tidyverse,magrittr,data.table)
}
